Adding stencil to the Frame Buffer:  DxGLFrameBuffer.cs
This small modification of the frame buffer allows the stencil tests to be used. That is, in certain cases, extremely fast and useful feature. It allows for example to render an area of nonconvex polygons (even with holes inside) knowing just their boundary. And much more.
This simple change has been originally suggested by Varon. I have made that change and tested on my code - works well.