Bypass Intel Driver bugs in driver v32.0.101.6557 (see remarks)
This commit circumvents two errors with Intel drivers:

1) A D3D9 surface cannot be registered as a Renderbuffer target
2) A D3D9 surface cannot be created with MSAA enabled

The Intel drivers have a bug where a D3D9 surface registered as
a Renderbuffer target type accesses memory out of range. There are
a host of theories as to why this is, but this commit bypasses
using a Renderbuffer target in lieu of a Texture2D target.

For depth testing and MSAA to work as intended, the user must
create their own framebuffers with these attachments.